+++ src/AutoUpdateServiceSample/SampleAutoUpdateServiceHost.cs
@@ -17,15 +17,19 @@ namespace AutoUpdateServiceSample
 
 		public SampleAutoServiceUpdatableHost()
 		{
-			System.Diagnostics.Trace.Listeners.Add(new System.Diagnostics.ConsoleTraceListener());
 			System.Diagnostics.Trace.AutoFlush = true;
+			var console = new System.Diagnostics.ConsoleTraceListener();
+			if (!System.Diagnostics.Trace.Listeners.Contains(console))
+			{
+				System.Diagnostics.Trace.Listeners.Add(console);
+			}
 
 			System.Diagnostics.Trace.WriteLine("Constructed");
 		}
 
 		public void Initialize()
 		{
-			System.Diagnostics.Trace.WriteLine("Autoupdate17 Initialized");
+			System.Diagnostics.Trace.WriteLine("Autoupdate18 Initialized");
 		}
 
 		public void Start()
@@ -34,7 +38,7 @@ namespace AutoUpdateServiceSample
 			m_Timer.Interval = 2 * 1000;
 			m_Timer.Elapsed += (s, arg) =>
 			{
-				System.Diagnostics.Trace.WriteLine(string.Format("Autoupdate17 Sample : {0}", DateTime.Now));
+				System.Diagnostics.Trace.WriteLine(string.Format("Autoupdate18 Sample : {0}", DateTime.Now));
 			};
 			m_Timer.Start();
 		}